+++ nightscout_secrets.py
@@ -1,0 +1,7 @@
+secrets = {
+    'ssid' : 'YOURSSID',
+    'password' : 'YOURPASSWORD',
+    'timezone' : "America/Los_Angeles", 
+    'token' : "api-fromyournightscoutsite",
+    'nightscout' : "https://YOURNIGHTSCOUTSITE"
+    }
