--- secrets.py
@@ -1,7 +1,0 @@
-secrets = {
-    'ssid' : 'YOURSSID',
-    'password' : 'YOURPASSWORD',
-    'timezone' : "America/Los_Angeles", 
-    'token' : "api-fromyournightscoutsite",
-    'nightscout' : "https://YOURNIGHTSCOUTSITE"
-    }
